+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ CHANGELOG
 UNRELEASED
 ----------
 
+* 🎉 feat: header component added
 * 🎉 feat: tailwind added
 
 .. 1.0.0 (yyyy-mm-dd)